Move files to /components and delete unnecessary files
