--- src/components/Temp.tsx
@@ -1,14 +1,0 @@
-import "../App.css";
-
-function Temp() {
-  return (
-    <div>
-      <main>
-        <p>Temp element until real pages are created</p>
-        <p>Will update as components are made</p>
-      </main>
-    </div>
-  );
-}
-
-export default Temp;
