+++ src/components/createAccount.tsx
@@ -2,8 +2,8 @@
 import { useState } from "react";
 import { useNavigate, Link } from "react-router-dom";
 import styled from "styled-components";
-import eyeSlash from "./images/EyeSlash.png";
-import backArrow from "./images/Back Arrow.png";
+import eyeSlash from "../images/EyeSlash.png";
+import backArrow from "../images/Back Arrow.png";
 
 const Wrapper = styled.section`
   display: flex;
