+++ src/components/success.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import styled from "styled-components";
-import logo from "./images/logo1.svg";
+import logo from "../images/logo1.svg";
 
 const Wrapper = styled.section`
   display: flex;
