+++ src/components/weeklyView.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from "react";
 import styled from "styled-components";
 // import { ChevronLeftIcon, ChevronRightIcon } from "@heroicons/react/solid";
-import chevronLeft from "./images/chevron left.svg";
+import chevronLeft from "../images/chevron left.svg";
 
 interface ICalendarWeekProps {
   startDate: Date;
