+++ src/components/weeklyViewMobile.tsx
