--- src/forgotPassword.css
@@ -1,110 +1,0 @@
-/* .fgCard {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-} */
-/* 
-.fgBox {
-  margin-top: -10%;
-  flex-direction: column;
-  /* width: 640px; */
-/* width: 45%;
-  height: auto;
-  border: 1px solid #c4c4c4;
-  padding: 0%;
-  align-content: stretch;
-  align-items: center; */
-/* margin-bottom: 10%; } */
-
-/* .insidefpPanel {
-  display: flex;
-  /* align-items: center; */
-/* justify-content: center;
-  flex-direction: column;
-  align-content: stretch;
-  padding: 10%;
-  padding-left: 15%;
-  padding-bottom: 15%; }/* */
-
-/* .fgText {
-  display: flex;
-  padding-bottom: 5%;
-  cursor: text;
-
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: 500;
-  /* font-size: 40px; */
-/* font-size: 280%;
-  /* line-height: 47px; */
-/* line-height: 100%; */
-/* color: #011338; }*/
-
-/* .fgBodyText {
-  font-family: "Rubik";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 110%;
-  /* line-height: 27px; */
-/* line-height: 150%; */
-/* or 135% */
-/* color: #000d26; }*/
-
-/* 
-.sendButton {
-  align-self: center;
-  /* login button font */
-/* font-family: "Rubik";
-  font-style: normal;
-  font-weight: 500;
-  font-size: 100%;
-  line-height: 300%;
-  text-align: center;
-  color: #ffffff; */
-/* button size*/
-/* width: 92%;
-  height: 100%;
-  /* Brand color 1 */
-/* background: #1b4c5a; */
-/* cursor: pointer; }*/
-
-/* 
-.textInput {
-  padding-top: 5%;
-  padding-bottom: 5%;
-  cursor: text;
-} */
-
-/* input {
-  align-self: left;
-  /* height: 55px;
-    width: 436px;
-    left: 538px; */
-/* line-height: 400%; */
-/* width: 90%; */
-/* left: 500%; }  */
-/* 
-
-h1 {
-  font-family: "Rubik";
-  font-style: normal;
-  font-weight: 500;
-  /* font-size: 18px;
-    line-height: 21px; */
-/* font-size: 112.5%; */
-/* line-height: 131.25%; */
-/* color: #011338; */
-/* } */
-
-/* .arrowButton {
-  scale: 25%;
-  margin-top: -5%;
-  margin-left: -6%;
-  outline: none;
-  background-color: white;
-  border: none;
-} */
-
-/* .invalidAlert {
-    border: red;
-} */
