--- src/home.tsx
@@ -1,3 +1,0 @@
-function Home() {}
-
-export default Home;
